Add compatibility for mir 0.31.0
This patch provides an alternative (and corrected) implementation to the
existing attempt in patch b0929330cfa625236366e1d9b0e6e5fc3313e924.

First and foremost, this patch makes the API change conditional on mir
0.31.0 (correct) instead of 0.30.0 (where no API change occurs).  There
are also a number of differences in the basic approach taken.

The changes required all resolve around the `SurfaceObserver` class,
whose methods now all expect to receive a pointer to the surface that
has changed.  This requires dependent code in qtmir to provide different
overrides and implementations depending on which server API version is
in use, which can be checked using `#if MIR_SERVER_VERSION ...`.

Placement of the required `#if` blocks has been designed to avoid nested
conditions and to minimize code duplication.  This means that:

  * for method declarations that were already conditional on the server
    API version, the existing `#if` checks have been extended into
    `#if`/`#elif` blocks

  * for method declarations not previously conditional, one big
    `#if`/`#else` block has been used

  * where implementations are concerned, separate `#if`/`#else` blocks
    are used for each method (to avoid co-dependencies); to avoid code
    duplication, where possible they have been placed only around the
    function declaration and not around the function body

  * this latter 'rule' has been ignored for a couple of functions with
    empty bodies whose implementation was already dependent on the API
    version: here the whole implementation has been placed inside the
    conditional blocks, since this allows for simpler code.

This approach is designed to bring greater clarity over which methods
and implementations are to be used with which mir versions, and should
allow easier maintenance in future.

Part of https://github.com/ubports/qtmir/issues/2.