Port to new mg::gl::Texture platform API.
This pulls out the existing QSGTexture implementation and replaces it with a
QSGMaterial/QSGMaterialShader implementation.

This does *not* cache the buffers received from Mir, so at the moment if
a MirSurfaceItem appears twice in the scenegraph the client will be able
to render at double speed. This should be fixed by updating the Mir
interfaces.