Move and Orientation impl out of screen and into it's own improved class
This moves the Orientation impl out of the screen impl where it listen for a
QSensor event once insted of on each screen class.

It also makes sure that screen runs on the orientations thread, this is needed
for signals to work.

This fixes: https://github.com/ubports/ubuntu-touch/issues/1063