Revert "Remove need for cgmanager, we got no need for upstart cgroups anymore"
This reverts commit d5989c616f39bf14b508f509cfd8694678b650ae, but
retains the more recent changes to move content-hub to an optional
dependency.

Removing or forward-porting the cgmanager to systemd must be done to run
qtmir on a systemd+ubuntu-app-launch4 system. xenial uses old
ubuntu-app-launch and upstart, though, so it needs to be able to
determine PIDs correctly.

Fixes https://github.com/ubports/ubuntu-touch/issues/1192