Terminate application while force closing surface
Not-well-behaved applications may ignore the close request. If we
proceed to force-close their surface and they do something crass like
request focus, we'll attempt to refocus the surface that was deleted.
This causes a use-after-free condition.

It would still be preferable to ask the user if they'd like to
force-close the application in Lomiri, but the swipe-away semantic in
the Spread makes that difficult.

Fixes https://github.com/ubports/ubuntu-touch/issues/1417