mirserver: uses UbuntuTheme from qtubuntu
There's a discrepancy between the icon theme used in qtmir (i.e. Unity8/
Lomiri) vs. the one used in qtubuntu (i.e. applications) and ubuntu-ui-
toolkit (used for `image://theme/` uri).

This commit copies the UbuntuTheme file from qtubuntu, which also brings
in the font settings. This is intended to be a temporary solution; the
better solution would be to let soon-to-be qtlomiri-appmenutheme expose
this theme and re-rename it to qtlomiri-themes [1].

This fixes the indicator-sound's microphone icons issue. ubuntu-mobile
icon theme doesn't have the icon the indicator expects, causing it to
failback to another icon, despite the fact that the icon will come from
Suru anyway.

[1] Or, better yet, maybe we shouldn't hardcode these things in and uses
some other mechanism to allow these to be configurable? I don't know.